+++ src/main/kotlin/org/nextras/orm/intellij/reference/ModifierClassPropertyProvider.kt
@@ -1,12 +1,9 @@
 package org.nextras.orm.intellij.reference
 
-import com.intellij.patterns.PlatformPatterns
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReference
 import com.intellij.psi.PsiReferenceProvider
 import com.intellij.util.ProcessingContext
-import com.jetbrains.php.lang.PhpLanguage
-import com.jetbrains.php.lang.documentation.phpdoc.lexer.PhpDocTokenTypes
 import com.jetbrains.php.lang.documentation.phpdoc.psi.PhpDocToken
 
 /**
@@ -14,17 +11,6 @@ import com.jetbrains.php.lang.documentation.phpdoc.psi.PhpDocToken
  * This make it click-able to the member definition.
  */
 class ModifierClassPropertyProvider : PsiReferenceProvider() {
-	companion object {
-		val PATTERN = PlatformPatterns
-			.psiElement(PhpDocTokenTypes.DOC_VARIABLE)
-			.afterLeaf(
-				PlatformPatterns.psiElement(PhpDocTokenTypes.DOC_STATIC)
-					.withText("::")
-					.afterLeaf(ModifierClassNameProvider.PATTERN)
-			)
-			.withLanguage(PhpLanguage.INSTANCE)
-	}
-
 	override fun getReferencesByElement(
 		element: PsiElement,
 		context: ProcessingContext