+++ changelog.md
@@ -2,9 +2,10 @@
 
 ## Unreleased
 
-- Fixed autocompletion & reference contribution (ctrl+click) in `getBy()` methods on various complex expression. #106
-
 ### Changed
+
+- Added autocompletion for class names in modifiers `{wrapper }`, `{m:1 }`, `{1:m}`, `{m:m }`, `{1:1 }`.
+- Fixed autocompletion & reference contribution (ctrl+click) in `getBy()` methods on various complex expression. #106
 
 ## 2.0.1 - 2025-01-22